adding auth POO

+++ show-article.php
@@ -1,5 +1,7 @@
 <?php
-
+require_once __DIR__.'/database/database.php';
+$authDb = require_once __DIR__.'/database/security.php';
+$currentUser = $authDb->isLoggedin();
 $articleDB = require_once __DIR__ . '/database/models/ArticleDB.php';
 $_GET = filter_input_array(INPUT_GET, FILTER_SANITIZE_FULL_SPECIAL_CHARS);
 $id = $_GET['id'] ?? '';
@@ -31,10 +33,13 @@ if (!$id) {
                 <h1 class="article-title"><?= $article['title'] ?></h1>
                 <div class="separator"></div>
                 <p class="article-content"><?= $article['content'] ?></p>
-                <div class="action">
-                    <a class="btn btn-secondary" href="/delete-article.php?id=<?= $article['id'] ?>">Supprimer</a>
-                    <a class="btn btn-primary" href="/form-article.php?id=<?= $article['id'] ?>">Editer l'article</a>
-                </div>
+                <p class="article-author"><?= $article["firstname"]." ".$article["lastname"] ?></p> 
+                <?php if($currentUser && $currentUser['id'] === $article['author']): ?>
+                    <div class="action">
+                        <a class="btn btn-secondary" href="/delete-article.php?id=<?= $article['id'] ?>">Supprimer</a>
+                        <a class="btn btn-primary" href="/form-article.php?id=<?= $article['id'] ?>">Editer l'article</a>
+                    </div>
+                <?php endif ?>    
             </div>
         </div>
         <?php require_once 'includes/footer.php' ?>
